20240802 last

+++ server/src/main/java/today/wtfood/server/dao/InquiryRepository.java
@@ -3,6 +3,6 @@ package today.wtfood.server.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import today.wtfood.server.entity.Inquiry;
 
-public interface InquiryRepository extends JpaRepository<Inquiry, Integer> {
+public interface InquiryRepository extends JpaRepository<Inquiry, Long> {
 
 }
